+++ src/CommandLine.h
@@ -46,6 +46,7 @@ public:
 	bool analyzeVsmCostJson() const;
 	bool compareVsmCost() const;
 	bool compareVsmCostJson() const;
+	bool compareVsmCostMarkdown() const;
 	bool dumpInstructionInfo() const;
 	bool dumpInstructionInfoJson() const;
 	const std::vector< std::pair<std::string, unsigned int> >& costLoops() const;
@@ -108,6 +109,7 @@ private:
 		ANALYZE_VSM_COST_LOOP,
 		ANALYZE_VSM_COST_COMPARE,
 		ANALYZE_VSM_COST_COMPARE_JSON,
+		ANALYZE_VSM_COST_COMPARE_MARKDOWN,
 		DUMP_INSTRUCTION_INFO,
 		DUMP_INSTRUCTION_INFO_JSON,
 
@@ -151,6 +153,7 @@ private:
 	bool m_analyzeVsmCostJson;
 	bool m_compareVsmCost;
 	bool m_compareVsmCostJson;
+	bool m_compareVsmCostMarkdown;
 	bool m_dumpInstructionInfo;
 	bool m_dumpInstructionInfoJson;
 	std::vector< std::pair<std::string, unsigned int> > m_costLoops;